add sitemap router, getSitemap function. call getSitemap() to generate sitemap.xml when server start
